refactor(common): move queryKeySerialize to presentation layer
queryKeySerialize is a presentation-only concern (toast dedupe key from a query key); every caller already lives in presentation, so it shouldn't cross the infrastructure boundary.
