+++ src/module/user/presentation/hooks/use-session-user.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
-import { queryKeySerialize } from '@/common/infrastructure/tanstack/query-key';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
+import { queryKeySerialize } from '@/common/presentation/tanstack/query-key';
 import { getSessionUserQueryOptions } from '@/user/infrastructure/tanstack/query/options';
 
 export function useSessionUser() {
